@@ -12,7 +12,7 @@ Devise.setup do |config|
   # Configure the e-mail address which will be shown in Devise::Mailer,
   # note that it will be overwritten if you use your own mailer class
   # with default "from" parameter.
-  config.mailer_sender = 'do-not-reply@carpe.us'
+  config.mailer_sender = 'Carpe <do-not-reply@carpe.us>'
 
   # Configure the class responsible to send e-mails.
   # config.mailer = 'Devise::Mailer'
@@ -239,10 +239,10 @@ Devise.setup do |config|
   # Add a new OmniAuth provider. Check the wiki for more information on setting
   # up on your models and hooks.
   # config.omniauth :github, 'APP_ID', 'APP_SECRET', scope: 'user,public_repo'
-	
+
 	require 'omniauth-google-oauth2'
 	config.omniauth :google_oauth2, "691236853053-d74vecl5lkc73seoi1qhgta87irllcak.apps.googleusercontent.com", "eK9EIkCjW0Gw5RLX8vbwgp0F", { access_type: "offline", approval_prompt: "" }
-	
+
   # ==> Warden configuration
   # If you want to use other strategies, that are not supported by Devise, or
   # change the failure app, you can configure them inside the config.warden block.
