@@ -1,8 +1,8 @@
 # To run all tests, in the project directory run the command:
-# bundle exec rake test
+# bundle exec rails test
 # ----------------------------------------
 # To run this test, in the project directory run the command:
-# bundle exec rake test test/controllers/users_controller_test.rb
+# bundle exec rails test test/controllers/users_controller_test.rb
 
 require 'test_helper'
 
@@ -19,17 +19,17 @@ class UsersControllerTest < ActionController::TestCase
 
   test "signed in user should be able to view other user" do
     sign_in @viktor
-    get :show, id: @norm.id
+    get :show, params: { id: @norm.id }
     assert_response :success
   end
 
   test "non-signed in user should be able to view user" do
-    get :show, id: @norm.id
+    get :show, params: { id: @norm.id }
     assert_response :success
   end
 
   test "user views can be accessed through their custom urls" do
-    get :show, id: @putin.custom_url
+    get :show, params: { id: @putin.custom_url }
     assert_response :success
   end
 
@@ -42,7 +42,7 @@ class UsersControllerTest < ActionController::TestCase
   end
 
   test "routes to a user's id should redirect to their custom url when present" do
-    get :show, id: @viktor.id
+    get :show, params: { id: @viktor.id }
     assert_redirected_to user_path(@viktor)
   end
 
@@ -60,27 +60,27 @@ class UsersControllerTest < ActionController::TestCase
 
   test "signed in users can view their own profile" do
     sign_in @viktor
-    get :show, id: @viktor
+    get :show, params: { id: @viktor }
     assert_select "div#profile-header"
   end
 
   test "users can navigate to the schedule tab" do
-    get :show, id: @norm.id, page: "schedule"
+    get :show, params: { id: @norm.id, page: "schedule" }
     assert_response :success
   end
 
   test "users can navigate to the followers tab" do
-    get :show, id: @norm.id, page: "followers"
+    get :show, params: { id: @norm.id, page: "followers" }
     assert_response :success
   end
 
   test "users can navigate to the following tab" do
-    get :show, id: @norm.id, page: "following"
+    get :show, params: { id: @norm.id, page: "following" }
     assert_response :success
   end
 
   test "users can navigate to the activity tab" do
-    get :show, id: @norm.id, page: "activity"
+    get :show, params: { id: @norm.id, page: "activity" }
     assert_response :success
   end
 
@@ -91,65 +91,65 @@ class UsersControllerTest < ActionController::TestCase
   # end
 
   test "users will navigate to the schedule tab by default" do
-    get :show, id: @norm.id
+    get :show, params: { id: @norm.id }
     assert_response :success
   end
 
   test "trying to view users that do not exist should redirect to the 404 page" do
-    get :show, id: "01010101010101"
+    get :show, params: { id: "01010101010101" }
     assert_response :missing
   end
 
   test "can perform an empty search query" do
-    get :search, q: nil
+    get :search, params: { q: nil }
     assert_response :success, "Accepts empty search queries"
   end
 
   test "can perform a search query" do
-    get :search, q: "v"
+    get :search, params: { q: "v" }
     assert_response :success
   end
 
   test "can perform a `json` search query" do
-    get :search, q: "v", format: "json"
+    get :search, params: { q: "v", format: "json" }
     assert_response :success
   end
 
   test "only admins (or the account owner) can delete users" do
     sign_in @norm
     assert_no_difference 'User.count' do
-      delete :destroy, id: @viktor
+      delete :destroy, params: { id: @viktor }
     end
 
     sign_out @norm
 
     sign_in @viktor
     assert_difference 'User.count', -1 do
-      delete :destroy, id: @norm
+      delete :destroy, params: { id: @norm }
     end
   end
 
   test "only admins can promote users" do
     sign_in @norm
-    get :promote, id: @putin
+    get :promote, params: { id: @putin }
     assert_not User.find(@putin.id).admin, "non-admin successfully promoted a user"
 
     sign_out @norm
 
     sign_in @viktor
-    get :promote, id: @putin
+    get :promote, params: { id: @putin }
     assert User.find(@putin.id).admin, "admin was unable to promote user"
   end
 
   test "only admins can view user information" do
     sign_in @norm
-    get :inspect, id: @norm
+    get :inspect, params: { id: @norm }
     assert_response :redirect
 
     sign_out @norm
 
     sign_in @viktor
-    get :inspect, id: @norm
+    get :inspect, params: { id: @norm }
     assert_response :success
   end
 end
